BAMF.jl Documentation

This repository is currently serving as a vehicle to explore the Julia Language environment using a somewhat sophisticated code base. It implements the core Reversible Jump Markov Chain Monte Carlo (RJMCMC) algorithm from the paper:

Bayesian Multiple Emitter Fitting using Reversible Jump Markov Chain Monte Carlo. Mohamadreza Fazel, Michael J. Wester, Hanieh Mazloom-Farsibaf, Marjolein B. M. Meddens, Alexandra S. Eklund, Thomas Schlichthaerle, Florian Schueder, Ralf Jungmann & Keith A. Lidke. Scientific Reports, 2019. 9(1): p. 13791.

Documentation is rudimentary and should be considered only as proof of principle of the autodoc system at this stage.

BAMF.ArrayDD — Type

ArrayDD <: BAMFData

data type for direct detection (i.e. a single image at a camera).

BAMF.BAMFData — Type

BAMFData

BAMFData is an abstract type. Specific data types will inherit from BAMFData. The most common example is 'directdetection'

BAMF.BAMFState — Type

BAMFState

BAMFState is an abstract type. Specific state types will inherit from BAMFState. States hold the current state of the RJMCMC Chain (i.e. θ) and can also be used to define a true underlying state that generated the data.

BAMF.PSF — Type

PSF

PSF is an abstract type. Specific psf types will inherit from PSF

BAMF.PSF_airy2D — Type

PSF_airy2D <: PSF

Airy PSF is ν²/(4π)(2J₁(νr)/(νr))² where ν=πD/(λf)=2π*nₐ/λ

Note that the Gaussian approximation is σ = 0.42*π/ν

BAMF.RJStruct — Type

RJStruct

RJStruct holds the data to be analyzed, the priors, the PSF model, and the parameters used in the RJMCMC steps.

BAMF.StateFlatBg — Type

StateFlatBg<: BAMFState

StateFlatBg is an abstract type with childern that implement CPU and GPU variants

BAMF.StateFlatBg_CPU — Type

StateFlatBg_CPU(n::Int32, x::Vector{Float32}, y::Vector{Float32}, photons::Vector{Float32}, bg::Float32)

State of a model that has x,y positions, total integrated intensity and a flat background offset.

BAMF.calcintialstate — Method

calcintialstate(rjs::RJStruct)

calculates an intial state by finding the maximum of the data and places an emitter at that location with an intensity selected from the prior.

BAMF.clusterstates — Method

clusterstates(states::Vector{Any},n::Int32)

uses kmeans clustering group x,y locations into n clusters and returns a State_Results structure.

BAMF.genBAMFData — Method

genBAMFData(rjs::RJStruct)

generate an empty BAMFData stucture for data type in RJStruct

BAMF.getmapnstates — Method

states_mapn=getmapn(states::Vector{Any})

returns a vector of states that where states.n == map_n

BAMF.getxy — Method

x,y,photons=getxy(states::Vector{Any})

returns a vector of all x,y,photons over all states

BAMF.likelihoodratio — Method

likelihoodratio(m::Array{Float32}, mtest::Array{Float32}, d::Array{Float32})

Generic Likelihood calculation using Poisson noise statistics

BAMF.poissrnd! — Method

" poissrnd!(d::Array{Float32}) In-place Poisson noise corruptor

BAMF.priorpdf — Method

" priorpdf(rjp::RJPrior,θ)

Calculate PDF(θ)

BAMF.randID — Method

randID(k::Int32)

return random number between 1 and k